#34 add  - excel list is delete api  - use delete mapping, patch mapping

+++ src/main/kotlin/net/dv/tax/domain/common/AccountingDataEntity.kt
@@ -16,7 +16,7 @@ data class AccountingDataEntity(
     @Id
     @Column(name = "id", insertable = false, updatable = false)
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    var id: Int? = null,
+    var id: Long? = null,
 
     var hospitalId: String? = null,
